chore(len_without_is_empty): extract to a separate module
It didn't share any logic at all with the lints in `len_zero.rs`,
whereas the helper functions were partially mixed together, so that the
file was hard to navigate
